+++ packages/angular-table/src/index.ts
@@ -14,10 +14,12 @@ export * from '@tanstack/table-core'
 
 export {
   type FlexRenderContent,
-  FlexRenderComponent,
   FlexRenderDirective,
   injectFlexRenderContext,
+  type FlexRenderComponentProps,
 } from './flex-render'
+
+export { FlexRenderComponent } from './flex-render/flex-render-component'
 
 export function createAngularTable<TData extends RowData>(
   options: () => TableOptions<TData>
